bet_maker/requirements.txt
aio-pika==9.5.4
aioredis==2.0.1
aiormq==6.8.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
black==24.10.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
pamqp==3.3.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.2.1
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
StrEnum==0.4.15
tomli==2.2.1
typing_extensions==4.12.2
uvicorn==0.34.0
yarl==1.18.3
